requirements.txt
Django==5.0.6
python-telegram-bot==21.3
requests==2.32.3
psycopg2-binary==2.9.9
dj-database-url==2.2.0
whitenoise==6.7.0
uvicorn==0.30.1
gunicorn==22.0.0